bpf: tests: condense LXC nodeport test
Run the IPv4 and IPv6 test from a single file.
